+++ tests/Unit/Api/IssueRelation/ListTest.php
@@ -1,0 +1,79 @@
+<?php
+
+namespace Redmine\Tests\Unit\Api\IssueRelation;
+
+use PHPUnit\Framework\TestCase;
+use Redmine\Api\IssueRelation;
+use Redmine\Client\Client;
+
+/**
+ * Tests for IssueRelation::list()
+ */
+class ListTest extends TestCase
+{
+    /**
+     * @covers \Redmine\Api\IssueRelation::list
+     */
+    public function testListWithoutParametersReturnsResponse()
+    {
+        // Test values
+        $response = '["API Response"]';
+        $expectedReturn = ['API Response'];
+
+        // Create the used mock objects
+        $client = $this->createMock(Client::class);
+        $client->expects($this->once())
+            ->method('requestGet')
+            ->with(
+                $this->stringStartsWith('/issues/5/relations.json')
+            )
+            ->willReturn(true);
+        $client->expects($this->exactly(1))
+            ->method('getLastResponseBody')
+            ->willReturn($response);
+        $client->expects($this->exactly(1))
+            ->method('getLastResponseContentType')
+            ->willReturn('application/json');
+
+        // Create the object under test
+        $api = new IssueRelation($client);
+
+        // Perform the tests
+        $this->assertSame($expectedReturn, $api->list(5));
+    }
+
+    /**
+     * @covers \Redmine\Api\IssuePriority::list
+     */
+    public function testListWithParametersReturnsResponse()
+    {
+        // Test values
+        $parameters = ['not-used'];
+        $response = '["API Response"]';
+        $expectedReturn = ['API Response'];
+
+        // Create the used mock objects
+        $client = $this->createMock(Client::class);
+        $client->expects($this->once())
+            ->method('requestGet')
+            ->with(
+                $this->logicalAnd(
+                    $this->stringStartsWith('/issues/5/relations.json'),
+                    $this->stringContains('not-used')
+                )
+            )
+            ->willReturn(true);
+        $client->expects($this->exactly(1))
+            ->method('getLastResponseBody')
+            ->willReturn($response);
+        $client->expects($this->exactly(1))
+            ->method('getLastResponseContentType')
+            ->willReturn('application/json');
+
+        // Create the object under test
+        $api = new IssueRelation($client);
+
+        // Perform the tests
+        $this->assertSame($expectedReturn, $api->list(5, $parameters));
+    }
+}